openapi/requirements.txt
Django==4.0.6
django-redis==5.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
huey==2.4.3
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
redis==4.3.4
